VR Update
I got the other coaster compressed and working as well as adding the
ability to change from narrow vr to stereographic 3d with a google
cardboard or a dive. Big todo list is to get the narrow vr working on
android which i’ll be able todo when my tango arrives on Wednesday and
i can troubleshoot properly.